Implement core simulation engine with queue management and metrics tracking
